Version 1.0.2 Release
This is the Release Version for Virtual Universe 1.0.2

Changes:

- Coverity fixes
- Code cleanup
- Code updates
- Updates from upstream providers

Important Note 1: Users and Grid Operators using Virtual Universe are strongly encourage to start with a fresh database and fresh copy of Virtual Universe to avoid false errors.

Important Note 2: When opening an issue please use the built in issue template provided.  By answering the questions you help us with the information we need to reproduce the bug or issue you are reporting.  This allows us to find and fix problems quickly.

+++ Universe/Physics/OpenDynamicsEngine/ODEDynamics.cs
@@ -1032,11 +1032,11 @@ namespace Universe.Physics.OpenDynamicsEngine
 
             #region deflection
             //rotates body to direction of movement (linearMovement vector)
-			/* temporary disabled due to instabilities, needs to be rewritten
+			// temporary disabled due to instabilities, needs to be rewritten
             if(m_angularDeflectionTimescale < 300)
             {
                 float Dservo = 0.05f * m_angularDeflectionTimescale * m_angularDeflectionEfficiency;
-                float mag = (float)linearVelocity.LengthSquared();
+                float mag = linearVelocity.LengthSquared();
                 if(mag > 0.01f) {
                     linearVelocity.Y = -linearVelocity.Y;
                     linearVelocity *= rotq;
@@ -1052,7 +1052,7 @@ namespace Universe.Physics.OpenDynamicsEngine
                     deflection *= Dservo;
                 }
             }
-			*/
+			//
             #endregion
 
             #region banking
